third_party/WebKit/LayoutTests/intersection-observer/iframe-no-root-expected.txt

Simple intersection observer test with no explicit root and target in an iframe.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS entries.length is 0
PASS entries.length is 0
PASS entries.length is 1
PASS entries[0].boundingClientRect.left is 18
PASS entries[0].boundingClientRect.right is 118
PASS entries[0].boundingClientRect.top is 468
PASS entries[0].boundingClientRect.bottom is 568
PASS entries[0].intersectionRect.left is 18
PASS entries[0].intersectionRect.right is 118
PASS entries[0].intersectionRect.top is 510
PASS entries[0].intersectionRect.bottom is 568
PASS entries[0].rootBounds.left is 0
PASS entries[0].rootBounds.right is 785
PASS entries[0].rootBounds.top is 0
PASS entries[0].rootBounds.bottom is 600
PASS entries[0].target is [object HTMLDivElement]
PASS entries.length is 2
PASS entries[1].boundingClientRect.left is 18
PASS entries[1].boundingClientRect.right is 118
PASS entries[1].boundingClientRect.top is 568
PASS entries[1].boundingClientRect.bottom is 668
PASS entries[1].intersectionRect.left is 0
PASS entries[1].intersectionRect.right is 0
PASS entries[1].intersectionRect.top is 0
PASS entries[1].intersectionRect.bottom is 0
PASS entries[1].rootBounds.left is 0
PASS entries[1].rootBounds.right is 785
PASS entries[1].rootBounds.top is 0
PASS entries[1].rootBounds.bottom is 600
PASS entries[1].target is [object HTMLDivElement]
PASS successfullyParsed is true

TEST COMPLETE
